Configures default theme and improves UI
Sets the default theme to 'simple' in blog configuration.

Updates the GitBook theme configuration to enable the map widget by default.

Refactors the Landing theme's Footer component by removing the copyright notice.

Adds DarkModeButton in simple theme header.

Removes and recreates the Footer component in the Simple theme. The content of the old Footer.js file was incomplete and not used, so it's safe to remove and add an empty one,the origin footer file should be in the correct directory.